+++ src/hkimjp/konpy2/routes.clj
@@ -19,8 +19,8 @@
     ["dl/:eid" {:get answers/dl}]
     ["logout" logout!]
     ["help"   {:get help}]
-    ["stats" {:get {:middleware [[m/wrap-admin] wrap-stats]
-                    :handler (fn [_] {:status 200 :body ""})}}]]
+    #_["stats" {:get {:middleware [m/wrap-admin wrap-stats]
+                      :handler (fn [_] {:status 200 :body ""})}}]]
    ["/admin/" {:middleware [m/wrap-admin]}
     [""           {:get admin/admin}]
     ["new"        {:get admin/new  :post admin/upsert!}]
@@ -59,7 +59,8 @@
       :not-acceptable
       (constantly {:status 406
                    :body "not acceptable"})}))
-   {:middleware [[wrap-defaults site-defaults]]}))
+   {:middleware [wrap-stats
+                 [wrap-defaults site-defaults]]}))
 
 ; (root-handler {:uri "/admin/eid" :request-method "post"})
 ; site-defaults